update dir name Room to room
