--- src/main/java/com/ssafy/ssafy_project/Room/adapter/in/web/dto/request/UpdateRoomRequest.java
@@ -1,6 +1,0 @@
-package com.ssafy.ssafy_project.Room.adapter.in.web.dto.request;
-
-public record UpdateRoomRequest(
-        String title
-) {
-}
