--- src/main/java/com/ssafy/ssafy_project/Room/adapter/in/web/dto/response/CreateRoomResponse.java
@@ -1,13 +1,0 @@
-package com.ssafy.ssafy_project.Room.adapter.in.web.dto.response;
-
-import java.time.LocalDateTime;
-
-public record CreateRoomResponse(
-        Long roomId,
-        String title,
-        Long hostId,
-        String roomCode,
-        LocalDateTime createdAt
-) {
-
-}
