--- src/main/java/com/ssafy/ssafy_project/Room/adapter/in/web/dto/response/UpdateRoomResponse.java
@@ -1,8 +1,0 @@
-package com.ssafy.ssafy_project.Room.adapter.in.web.dto.response;
-
-public record UpdateRoomResponse(
-        Long roomId,
-        String title
-) {
-
-}
