--- src/main/java/com/ssafy/ssafy_project/Room/application/port/in/CreateRoomCommand.java
@@ -1,7 +1,0 @@
-package com.ssafy.ssafy_project.Room.application.port.in;
-
-public record CreateRoomCommand(
-        String title,
-        Long hostId
-) {
-}
