--- src/main/java/com/ssafy/ssafy_project/Room/application/port/in/CreateRoomPortIn.java
@@ -1,5 +1,0 @@
-package com.ssafy.ssafy_project.Room.application.port.in;
-
-public interface CreateRoomPortIn {
-    CreateRoomResult createRoom(CreateRoomCommand createRoomCommand);
-}
