--- src/main/java/com/ssafy/ssafy_project/Room/application/port/in/CreateRoomResult.java
@@ -1,12 +1,0 @@
-package com.ssafy.ssafy_project.Room.application.port.in;
-
-import java.time.LocalDateTime;
-
-public record CreateRoomResult(
-        Long roomId,
-        String title,
-        Long hostId,
-        String roomCode,
-        LocalDateTime createdAt
-) {
-}
