--- src/main/java/com/ssafy/ssafy_project/Room/application/port/in/DeleteRoomCommand.java
@@ -1,7 +1,0 @@
-package com.ssafy.ssafy_project.Room.application.port.in;
-
-public record DeleteRoomCommand(
-        Long roomId,
-        Long userId
-) {
-}
