--- src/main/java/com/ssafy/ssafy_project/Room/application/port/in/DeleteRoomPortIn.java
@@ -1,5 +1,0 @@
-package com.ssafy.ssafy_project.Room.application.port.in;
-
-public interface DeleteRoomPortIn {
-    void deleteRoom(DeleteRoomCommand deleteRoomCommand);
-}
