--- src/main/java/com/ssafy/ssafy_project/Room/application/port/in/UpdateRoomCommand.java
@@ -1,8 +1,0 @@
-package com.ssafy.ssafy_project.Room.application.port.in;
-
-public record UpdateRoomCommand(
-        Long roomId,
-        String title,
-        Long userId
-) {
-}
