+++ src/main/java/com/ssafy/ssafy_project/room/adapter/in/web/dto/request/CreateRoomRequest.java
@@ -1,0 +1,6 @@
+package com.ssafy.ssafy_project.room.adapter.in.web.dto.request;
+
+public record CreateRoomRequest(
+        String title
+) {
+}
