+++ src/main/java/com/ssafy/ssafy_project/room/adapter/in/web/dto/request/UpdateRoomRequest.java
@@ -1,0 +1,6 @@
+package com.ssafy.ssafy_project.room.adapter.in.web.dto.request;
+
+public record UpdateRoomRequest(
+        String title
+) {
+}
