+++ src/main/java/com/ssafy/ssafy_project/room/adapter/in/web/dto/response/UpdateRoomResponse.java
@@ -1,0 +1,8 @@
+package com.ssafy.ssafy_project.room.adapter.in.web.dto.response;
+
+public record UpdateRoomResponse(
+        Long roomId,
+        String title
+) {
+
+}
