+++ src/main/java/com/ssafy/ssafy_project/room/application/port/in/CreateRoomPortIn.java
@@ -1,0 +1,5 @@
+package com.ssafy.ssafy_project.room.application.port.in;
+
+public interface CreateRoomPortIn {
+    CreateRoomResult createRoom(CreateRoomCommand createRoomCommand);
+}
