+++ src/main/java/com/ssafy/ssafy_project/room/application/port/in/CreateRoomResult.java
@@ -1,0 +1,12 @@
+package com.ssafy.ssafy_project.room.application.port.in;
+
+import java.time.LocalDateTime;
+
+public record CreateRoomResult(
+        Long roomId,
+        String title,
+        Long hostId,
+        String roomCode,
+        LocalDateTime createdAt
+) {
+}
