+++ src/main/java/com/ssafy/ssafy_project/room/application/port/in/DeleteRoomCommand.java
@@ -1,0 +1,7 @@
+package com.ssafy.ssafy_project.room.application.port.in;
+
+public record DeleteRoomCommand(
+        Long roomId,
+        Long userId
+) {
+}
