+++ src/main/java/com/ssafy/ssafy_project/room/application/port/in/UpdateRoomCommand.java
@@ -1,0 +1,8 @@
+package com.ssafy.ssafy_project.room.application.port.in;
+
+public record UpdateRoomCommand(
+        Long roomId,
+        String title,
+        Long userId
+) {
+}
