+++ src/main/java/com/ssafy/ssafy_project/room/application/port/in/UpdateRoomPortIn.java
@@ -1,0 +1,5 @@
+package com.ssafy.ssafy_project.room.application.port.in;
+
+public interface UpdateRoomPortIn {
+    UpdateRoomResult updateRoom(UpdateRoomCommand updateRoomCommand);
+}
