+++ src/main/java/com/ssafy/ssafy_project/room/application/port/in/UpdateRoomResult.java
@@ -1,0 +1,7 @@
+package com.ssafy.ssafy_project.room.application.port.in;
+
+public record UpdateRoomResult(
+        Long roomId,
+        String title
+) {
+}
